requirements.txt
tqdm
numpy
matplotlib
scipy
torch
torchvision
torchtext
torchaudio
ipython
ipywidgets
IProgress
librosa
SoundFile
inflect
PyYAML
notebook
unidecode
webrtcvad-wheels